+++ CHANGES.md
@@ -13,6 +13,7 @@ Changes in 7.4
 - Do not display "General Comments" title if no comments in ReportCards.fnc.php
 - HTML display rows of 3 School Period checkboxes in AddAbsences.php
 - Comment input maxlength increased to 500 in InputFinalGrades.php
+- Comment Code input field is required in ReportCardCommentCodes.php
 
 Changes in 7.3.1
 ----------------